project2/requirements.txt
pyspark
kafka-python
cassandra-driver==3.26.0
python-dotenv